+++ client/src/App.js
@@ -3,10 +3,6 @@ import About from "./pages/About";
 import ModalTestingPage from "./pages/ModalTestingPage";
 import Home from "./pages/Home";
 import SliderTest from "./pages/SliderTest";
-import DataFetchingTest from "./pages/DataFetchingTest";
-
-import RhythmsPanelTestingPage from "./pages/RhythmsPanelTestingPage";
-
 
 
 const App = () => (
@@ -18,7 +14,6 @@ const App = () => (
 		<Route path="/" element={<Home />} />
 		<Route path="/about/this/site" element={<About />} />
 		<Route path="/testingpage" element={<ModalTestingPage />} />
-			<Route path="/RhythmsPanelTestingPage" element={<RhythmsPanelTestingPage />} />
 		</Routes>
 
 	</>
