+++ client/src/components/MyMap.css
@@ -1,0 +1,5 @@
+.map {
+	position: relative;
+	top: 0;
+	left: 0;
+}
