+++ client/src/components/RhythmCodeIcon.css
@@ -1,0 +1,4 @@
+.rhythmCodeIcon {
+	width: 40px;
+	height: 40px;
+}
