+++ client/src/components/RhythmItem.css
@@ -1,0 +1,32 @@
+.rhythmItemList {
+	position: absolute;
+	list-style-type: none;
+	padding: 15px;
+	background-color: rgba(255, 255, 255, 0.8);
+	border: 1px solid transparent;
+	border-radius: 10px;
+	width: 120px;
+	color: var(--copy-color);
+}
+
+.rhythmCodeName {
+	font-weight: 700;
+	list-style-type: none;
+	margin: 5px 0;
+	padding-bottom: 8px;
+}
+
+.rhythmItemList > li > button {
+	text-align: left;
+	background: none;
+	border: none;
+	padding: 1px;
+	font: var(--font-roboto);
+	color: var(--copy-color);
+	font-size: 16px;
+	cursor: pointer;
+}
+
+.rhythmItemList > li:hover {
+	font-weight: 700;
+}
