+++ client/src/components/RhythmItem.js
@@ -1,14 +1,12 @@
 function RhythmItem({ setModalOpen, setSelectedRhythm, rhythms }) {
-
 	const handleRhythmClick = () => {
 		setModalOpen(true);
 		setSelectedRhythm({ ...rhythms });
-		console.log("Clicked selected rhythmmmmmmm", `${rhythms.rhythm}`);
 	};
 	return (
-			<li>
-				<button onClick={handleRhythmClick}>{rhythms.rhythm}</button>
-			</li>
+		<li>
+			<button onClick={handleRhythmClick}>{rhythms.rhythm}</button>
+		</li>
 	);
 }
